+++ app/scripts/agents/run_agent_best.py
@@ -1,9 +1,8 @@
-#!/usr/bin/python3
-
-from os.path import dirname, realpath
-import yaml
-import os
 import argparse
+import os
+import yaml
+from os.path import dirname, realpath, sep, pardir
+os.chdir(dirname(realpath(__file__)) + sep + pardir + sep + pardir + sep)
 
 from irec.connector import utils
 import argparse
@@ -21,19 +20,16 @@ parser.add_argument("--agents", nargs="*", default=[settings["defaults"]["agent"
 
 parser.add_argument("--tasks", type=int, default=os.cpu_count())
 
-parser.add_argument("--metrics", nargs="*", default=[settings["defaults"]["metric"]])
-parser.add_argument(
-    "--metric_evaluator", default=settings["defaults"]["metric_evaluator"]
-)
 parser.add_argument("--forced_run", action='store_true', default=False)
+utils.load_settings_to_parser(settings, parser)
 args = parser.parse_args()
+settings = utils.sync_settings_from_args(settings, args)
 
+settings["defaults"]["evaluation_policy"] = args.evaluation_policy
 
 dataset_agents_parameters = yaml.load(
     open("./settings/dataset_agents.yaml"), Loader=yaml.SafeLoader
 )
 
-settings["defaults"]["evaluation_policy"] = args.evaluation_policy
-settings["defaults"]["metric_evaluator"] = args.metric_evaluator
+utils.run_agent_with_dataset_parameters(args.agents,args.dataset_loaders,settings,dataset_agents_parameters, args.tasks,args.forced_run)
 
-utils.evaluate_agent_with_dataset_parameters(args.agents,args.dataset_loaders,settings,dataset_agents_parameters, args.metrics, args.tasks,args.forced_run)
