+++ app/scripts/environment/generate_dataset.py
@@ -1,14 +1,9 @@
 #!/usr/bin/python3
 from os.path import dirname, realpath, sep, pardir
-import sys
-
-sys.path.append(dirname(realpath(__file__)) + sep + pardir)
-
-from app import constants
+import os
+os.chdir(dirname(realpath(__file__)) + sep + pardir + sep + pardir + sep)
 from irec.connector import utils
 import argparse
-from irec.utils.Factory import DatasetLoaderFactory
-import mlflow
 
 settings = utils.load_settings(dirname(realpath(__file__)))
 parser = argparse.ArgumentParser()
