+++ app/scripts/evaluation/eval_agent_search.py
@@ -1,12 +1,12 @@
-#!/usr/bin/python3
-
-from os.path import dirname, realpath
-import yaml
+from os.path import dirname, realpath, sep, pardir
 import os
+import yaml
+os.chdir(dirname(realpath(__file__)) + sep + pardir + sep + pardir + sep)
 import argparse
-from irec.connector.utils import load_dataset_experiment
+
 from irec.connector import utils
 import argparse
+
 from irec.recommendation.hyperoptimization.grid_search import GridSearch
 
 settings = utils.load_settings(dirname(realpath(__file__)))
@@ -20,7 +20,10 @@ parser.add_argument(
 )
 parser.add_argument("--agents", nargs="*", default=[settings["defaults"]["agent"]])
 parser.add_argument("--tasks", type=int, default=os.cpu_count())
-parser.add_argument("--forced_run", action='store_true', default=False)
+parser.add_argument("--metrics", nargs="*", default=[settings["defaults"]["metric"]])
+parser.add_argument(
+    "--metric_evaluator", default="CumulativeInteractionMetricEvaluator"
+)
 parser.add_argument("--tunning", default="GridSearch")
 
 utils.load_settings_to_parser(settings, parser)
@@ -33,12 +36,10 @@ agents_variables = [template for agent_name in args.agents for template in agent
 
 g = GridSearch()
 agents_search = g.generate_settings(agents_variables)
+
+
 settings["defaults"]["evaluation_policy"] = args.evaluation_policy
+settings["defaults"]["metric_evaluator"] = args.metric_evaluator
 
-for dataset_loader_name in args.dataset_loaders:
-    settings["defaults"]["dataset_loader"] = dataset_loader_name
-    train_dataset, test_dataset, x_validation, y_validation = load_dataset_experiment(settings, validation=True)
-    g.execute(x_validation, y_validation, agents_search, settings, args.tasks, args.forced_run)
-
-# utils.run_agent_search(args.agents,args.dataset_loaders,
-        # settings, agents_search, args.tasks,args.forced_run)
+utils.eval_agent_search(args.agents,args.dataset_loaders,
+        settings,agents_search,args.metrics, args.tasks)
