+++ app/scripts/others/acts_info_analysis.py
