+++ app/scripts/others/constants.py
