+++ app/scripts/others/copy_parameters.py
