+++ app/scripts/others/generate_search_parameters.py
