+++ app/scripts/others/parameters_generator.py
