+++ app/scripts/others/plot_hits_users_coverage.py
