+++ app/scripts/others/print_methods_info.py
