+++ app/scripts/others/run_search_rates.sh
