+++ app/scripts/others/settings_update_format.py
