+++ app/scripts/others/settings_update_format_search.py
