+++ app/settings/dataset_agents.yaml
@@ -4141,8 +4141,13 @@ MovieLens 100k:
         ASPGreedy: {}
       value_function:
         GLM_UCB:
-          c: 1.0
+          c: 4
+          item_var: 0.01
+          iterations: 20
           num_lat: 10
+          stop_criteria: 0.0009
+          user_var: 0.01
+          var: 0.05
   HELF:
     SimpleAgent:
       action_selection_policy:
